refactor: extract resource_ref_for into looplet.refs
Move the small resource-identity registry (_resource_origin,
_register_resource_origin, resource_ref_for, _REF_PREFIX) out of
the 3000-line workspace.py loader into a new tiny looplet.refs
module so core modules can call resource_ref_for without any
hard dependency on the workspace loader.

Before this change, four core modules (permissions, streaming,
evals, telemetry) each had:

    from looplet.workspace import resource_ref_for  # lazy local import

…inside their to_config() methods. The lazy import worked but
made the workspace loader an effective hard dep of core: anyone
extracting workspace.py into a separate package would have had
to either move resource_ref_for or keep a back-import.

After this change, the four core modules import from
looplet.refs (~120 lines, zero looplet deps). The workspace
loader still re-exports the same symbols so any existing test or
consumer that reaches looplet.workspace._resource_origin /
.resource_ref_for keeps working unchanged.

Verification: 1887 passed, 1 skipped; ruff + pyright clean.
